+++ packages/image/.babelrc.js
@@ -1,0 +1,15 @@
+const buildTarget = process.env.BUILD_TARGET || 'esmodule' // another value could be 'commonjs'
+
+export default {
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        modules: buildTarget === 'esmodule' ? false : 'commonjs',
+        targets: {
+          node: '14',
+        },
+      },
+    ],
+  ],
+}