+++ packages/image/src/index.js
@@ -1,3 +1,5 @@
-import Image from './react-components'
+import Image from './react-components/index.js'
+
+export * from './types/index.js'
 
 export default Image