+++ packages/timeline/src/react-components/index.js
@@ -1,5 +1,4 @@
 import styled, { createGlobalStyle } from 'styled-components'
-import TimelineUnit from './timeline-unit'
 import {
   calcNextLevelUnitKey,
   generateTimeLevel,
@@ -14,13 +13,38 @@ import TimelineEvent from './timeline-event'
 import { TagsContext, initialTags } from './useTags'
 import { defaultConifg } from '../const/config'
 import { useTimelineConfig } from './hook/useTimelineConfig'
+import TimelineList from './timeline-list'
+import useBlockPageScroll from './hook/useBlockPageScroll'
 
-const GlobalStyles = createGlobalStyle`
+const BlockBodyScrollGlobalStyle = createGlobalStyle`
+  body {
+    overflow: hidden;
+  }
+`
+
+const Wrapper = styled.div`
+  background-color: #efefef;
+  overflow: hidden;
+  width: 100vw;
+  height: calc(
+    100vh -
+      ${({ headerHeight }) => (headerHeight ? `${headerHeight}px` : '0px')}
+  );
+  position: relative;
+  left: calc(50% - 50vw);
+
+  ${({ isFixedMode, headerHeight }) =>
+    isFixedMode &&
+    `
+    position: fixed;
+    top: ${headerHeight}px;
+  `}
+
+  // reset for Timeline only
   * {
     box-sizing: border-box;
     font-family: Noto Sans CJK TC;
     -webkit-tap-highlight-color: transparent;
-
   }
 
   button {
@@ -52,14 +76,6 @@ const GlobalStyles = createGlobalStyle`
   }
 `
 
-const Wrapper = styled.div`
-  background-color: #efefef;
-  overflow: hidden;
-  width: 100vw;
-  position: relative;
-  left: calc(50% - 50vw);
-`
-
 const TimelineWrapper = styled.div`
   position: relative;
   width: 320px;
@@ -73,7 +89,22 @@ const TimelineWrapper = styled.div`
 `
 
 const TimelineNodesWrapper = styled.div`
+  // hide react-virtualized List scrollbar
+  &,
+  & > div {
+    -ms-overflow-style: none; /* IE and Edge */
+    scrollbar-width: none; /* Firefox */
+    &::-webkit-scrollbar {
+      display: none; /* for Chrome, Safari and Opera */
+    }
+  }
+  overflow: hidden auto;
+  position: relative;
+  width: 100vw;
+  left: calc(50% - 50vw);
+  ${({ height }) => height && `height: ${height}px;`}
   ${({ eventMode }) => eventMode && `padding: 0 36px 12px;`}
+  ${({ isFixedMode }) => !isFixedMode && `overflow: hidden !important;`}
 `
 
 const EventWrapper = styled.div`
@@ -85,11 +116,6 @@ const EventWrapper = styled.div`
     margin: 12px auto 0;
   }
 `
-
-function getBubbleLevel(max, count) {
-  const levelSize = max / 5
-  return Math.ceil(count / levelSize) - 1
-}
 
 /**
  * @param {Object} props
@@ -107,7 +133,6 @@ export default function Timeline({
     [liveblog, isTimeSortedAsc]
   )
   const [tags, setTags] = useState(initialTags)
-  const [stickyStrategy, setStickStrategy] = useState('absolute-top')
   const [focusUnitKey, setFocusUnitKey] = useState('')
 
   const timelineConfig = timeline.config || defaultConifg
@@ -136,18 +161,30 @@ export default function Timeline({
   const measure = getMeasureFromLevel(level)
   const timeUnitEvents = timeEvents[measure]
   const timeUnitKeys = timeKeys[measure]
-  const firstTimeUnitKey = timeUnitKeys[0]
-  const lastTimeeUnitKey = timeUnitKeys[timeUnitKeys.length - 1]
   const timeUnitKeysToRender = timeKeysToRender[measure]
+  const firstTimeUnitKeyToRender = timeUnitKeysToRender[0]
+  const lastTimeUnitKeyToRender =
+    timeUnitKeysToRender[timeUnitKeysToRender.length - 1]
+  const divider = dividers[measure]
+  const [listDimension, setListDimension] = useState({
+    width: 0,
+    height: 0,
+  })
+  const listItemHeight = listDimension.height / divider
+
   /** @type {React.RefObject<HTMLDivElement>} */
   const containerRef = useRef(null)
 
   const scroIntoViewType = useRef('')
   const blockOnScrollEvent = useRef(false)
-  /** @type {React.RefObject<HTMLDivElement>} */
-  const topRef = useRef(null)
-  /** @type {React.RefObject<HTMLDivElement>} */
-  const bottomRef = useRef(null)
+  const timelineListRef = useRef(null)
+
+  const { wrapperRef, isFixedMode } = useBlockPageScroll(
+    headerHeight,
+    measure !== 'event'
+      ? timelineListRef.current?.Grid._scrollingContainer
+      : containerRef.current
+  )
 
   const updateLevel = (newLevel, spFocusUnitKey) => {
     const oldFocusUnitKey = spFocusUnitKey || focusUnitKey
@@ -163,150 +200,180 @@ export default function Timeline({
   }
 
   useEffect(() => {
-    const onScroll = () => {
+    const resizeHandler = () => {
+      setListDimension({
+        width: window.innerWidth,
+        height: window.innerHeight - headerHeight,
+      })
+    }
+    window.addEventListener('resize', resizeHandler)
+    resizeHandler()
+    return () => {
+      window.removeEventListener('resize', resizeHandler)
+    }
+  }, [headerHeight])
+
+  // handle event mode updating focusUnitKey when user scrolling
+  useEffect(() => {
+    if (level !== 1) {
+      // only event mode needs onScroll event to update focusUnitKey
+      return
+    }
+
+    if (containerRef.current) {
+      const onScroll = () => {
+        const containerDiv = containerRef.current
+        // block update focusUnitKey when programmatically scrolling
+        if (blockOnScrollEvent.current) {
+          return
+        }
+        let newFocusNode
+        for (const node of containerDiv.children) {
+          const rect = node.getBoundingClientRect()
+          if (rect.top - headerHeight + rect.height > 0) {
+            newFocusNode = node
+            break
+          }
+        }
+        let newFocusUnitKey = newFocusNode.id.split('-')[1]
+        setFocusUnitKey(newFocusUnitKey)
+      }
+
+      containerRef.current.addEventListener('scroll', onScroll)
+
+      return () => {
+        containerRef.current.removeEventListener('scroll', onScroll)
+      }
+    }
+  }, [level, headerHeight])
+
+  // handle timeline mode updating focusUnitKey when user scrolling
+  const onTimelineListScroll = () => {
+    if (containerRef.current) {
       /** @type {HTMLDivElement} */
       const containerDiv = containerRef.current
       const containerTop = containerDiv.getBoundingClientRect().top
       if (blockOnScrollEvent.current) {
         return
       }
-      function getIndexOfTheTopMostItem(top, parentDom) {
-        if (top >= headerHeight) {
-          return 0
-        } else {
-          let topSum = top
-          for (let i = 0; i < parentDom.children.length; i++) {
-            let node = parentDom.children[i]
-            topSum += node.getBoundingClientRect().height
-            if (topSum >= headerHeight) {
-              return i
-            }
-            if (i === parentDom.children.length - 1) {
-              return parentDom.children.length - 1
-            }
-          }
+      const timeNodes = containerDiv.querySelectorAll('.timeline-list-item')
+      let firstVisibleIndex
+      for (let index = 0; index < timeNodes.length; index++) {
+        const node = timeNodes[index]
+        const distanceToTop =
+          node.getBoundingClientRect().top -
+          containerTop +
+          node.getBoundingClientRect().height -
+          10
+        if (distanceToTop > 0) {
+          firstVisibleIndex = index
+          break
         }
       }
-      const index = getIndexOfTheTopMostItem(containerTop, containerDiv)
-      let focusNode = containerDiv.children[index]
-      let focusUnitKey = focusNode.id.split('-')[1]
-      setFocusUnitKey(focusUnitKey)
+      let focusNode = timeNodes[firstVisibleIndex]
+      if (focusNode) {
+        let newFocusUnitKey = focusNode.firstChild.id.split('-')[1]
+        setFocusUnitKey(newFocusUnitKey)
+      } else {
+        // Use cmd up or down to control timeline could cause lost track of focus
+        // ignore it for now.
+        console.warn('onTimelineListScroll focusNode undefined', focusNode)
+      }
     }
-    window.addEventListener('scroll', onScroll)
+  }
 
-    return () => {
-      window.removeEventListener('scroll', onScroll)
-    }
-  }, [headerHeight])
-
+  // handle scroll timeline (List, Events) to right position
   useEffect(() => {
-    if (containerRef.current && scroIntoViewType.current) {
+    // block no scrollType
+    if (!scroIntoViewType.current) {
+      return
+    }
+
+    if (timelineListRef.current) {
+      // handle scrolling on react.virtualized List
+      const focusIndex = timeUnitKeysToRender.findIndex(
+        (unitKey) => unitKey === focusUnitKey
+      )
+
+      if (focusIndex !== -1) {
+        if (scroIntoViewType.current === 'immediate') {
+          // since scrollToRow can guarantee to scroll the focusItem to the topmost, use scrollToPosition instead
+          timelineListRef.current.scrollToPosition(
+            focusIndex * listItemHeight + 5
+          )
+        } else if (scroIntoViewType.current === 'smooth') {
+          function smoothScrollTo(targetPosition, speedFactor) {
+            const startPosition =
+              timelineListRef.current.Grid._scrollingContainer.scrollTop
+            const distance = Math.abs(targetPosition - startPosition)
+            const speed = speedFactor
+
+            const duration = Math.min(2000, Math.max(300, distance / speed)) // Set a maximum and minimum duration to avoid extremely long or short scrolls
+            const startTime = performance.now()
+
+            function scrollStep(timestamp) {
+              const currentTime = timestamp - startTime
+              const scrollFraction = currentTime / duration
+
+              if (currentTime >= duration) {
+                timelineListRef.current.scrollToPosition(targetPosition)
+                blockOnScrollEvent.current = false
+              } else {
+                const easeValue = scrollFraction ** 2 // You can adjust the easing function here
+                const scrollValue =
+                  startPosition +
+                  (targetPosition > startPosition ? 1 : -1) *
+                    distance *
+                    easeValue
+                timelineListRef.current.scrollToPosition(scrollValue)
+                window.requestAnimationFrame(scrollStep)
+                return
+              }
+            }
+
+            window.requestAnimationFrame(scrollStep)
+          }
+
+          blockOnScrollEvent.current = true
+          smoothScrollTo(focusIndex * listItemHeight + 5, 20)
+        }
+        scroIntoViewType.current = ''
+      } else {
+        // since focusKey can't map to a html node (may be filtered out)
+        // scroll to the first node
+        const newFocusUnitKey = timeUnitKeys[0]
+        setFocusUnitKey(newFocusUnitKey)
+        blockOnScrollEvent.current = true
+      }
+    } else if (containerRef.current) {
+      // handle scrolling on event mode
+
       const focusTimelineUnitEle = containerRef.current.querySelector(
         `#node-${focusUnitKey}`
       )
-      function smoothScrollTo(targetPosition, speedFactor) {
-        const startPosition = window.scrollY
-        const distance = Math.abs(targetPosition - startPosition)
-        const speed = speedFactor
 
-        const duration = Math.min(2000, Math.max(300, distance / speed)) // Set a maximum and minimum duration to avoid extremely long or short scrolls
-        const startTime = performance.now()
-
-        function scrollStep(timestamp) {
-          const currentTime = timestamp - startTime
-          const scrollFraction = currentTime / duration
-
-          if (currentTime >= duration) {
-            window.scrollTo(0, targetPosition)
-          } else {
-            const easeValue = scrollFraction ** 2 // You can adjust the easing function here
-            const scrollValue =
-              startPosition +
-              (targetPosition > startPosition ? 1 : -1) * distance * easeValue
-            window.scrollTo(0, scrollValue)
-            window.requestAnimationFrame(scrollStep)
-            return
-          }
-          blockOnScrollEvent.current = false
-        }
-
-        window.requestAnimationFrame(scrollStep)
-      }
       if (focusTimelineUnitEle) {
         // add 2 px to prevent focusIndex count on scroll mistaken
         if (scroIntoViewType.current === 'immediate') {
           blockOnScrollEvent.current = false
-          window.scrollTo(
+          containerRef.current.scrollTo(
             0,
-            window.scrollY +
+            containerRef.current.scrollTop +
               focusTimelineUnitEle.getBoundingClientRect().top +
-              2 -
+              5 -
               headerHeight
-          )
-        } else if (scroIntoViewType.current === 'smooth') {
-          blockOnScrollEvent.current = true
-          // add 2 px to prevent focusIndex count on scroll mistaken
-          smoothScrollTo(
-            window.scrollY +
-              focusTimelineUnitEle.getBoundingClientRect().top +
-              2 -
-              headerHeight,
-            20
           )
         }
         scroIntoViewType.current = ''
       } else {
         // since focusKey can't map to a html node (may be filtered out)
-        // find another valid key to trigger scroll again
-        const newFocusUnitKey = timeUnitKeys
-          .reverse()
-          .find((timeUnitKey) =>
-            isTimeSortedAsc
-              ? Number(focusUnitKey) > Number(timeUnitKey)
-              : Number(focusUnitKey) < Number(timeUnitKey)
-          )
+        // scroll to the first node
+        const newFocusUnitKey = timeUnitKeys[0]
         setFocusUnitKey(newFocusUnitKey)
         blockOnScrollEvent.current = true
       }
     }
   })
-
-  useEffect(() => {
-    if (topRef.current && bottomRef.current) {
-      const observer = new IntersectionObserver(
-        (entries) => {
-          entries.forEach(({ boundingClientRect }) => {
-            if (!boundingClientRect.width || !containerRef.current) {
-              return
-            }
-
-            const bounding = containerRef.current.getBoundingClientRect()
-            if (bounding.height < window.innerHeight) {
-              setStickStrategy('absolute')
-              return
-            }
-            if (bounding.height) {
-              if (bounding.y >= headerHeight) {
-                setStickStrategy('absolute-top')
-              } else if (bounding.y + bounding.height > window.innerHeight) {
-                setStickStrategy('fixed')
-              } else {
-                setStickStrategy('absolute-bottom')
-              }
-            }
-          })
-        },
-        { rootMargin: `-${headerHeight}px 0px 0px 0px` }
-      )
-
-      observer.observe(topRef.current)
-      observer.observe(bottomRef.current)
-
-      return () => {
-        observer.disconnect()
-      }
-    }
-  }, [level])
 
   const addTag = (newTag, timeUnitKey) => {
     if (window.screen.width < 768 && tags.length === 3) {
@@ -319,7 +386,7 @@ export default function Timeline({
       /*
        * since after filter there could be two condition:
        * 1. newFocusUnitKey still exist -> stay in the window
-       * 2. newFocusUnitKey got filtered -> find the cloest last oen to scroll
+       * 2. newFocusUnitKey got filtered -> scroll to the first one
        */
       setFocusUnitKey(newFocusUnitKey)
       scroIntoViewType.current = 'immediate'
@@ -337,54 +404,59 @@ export default function Timeline({
       : null
 
   let timelineNodesJsx =
-    measure !== 'event'
-      ? timeUnitKeysToRender.map((timeUnitKey, i) => {
-          const events = timeUnitEvents[timeUnitKey] || []
-          return (
-            <TimelineUnit
-              eventsCount={events.length}
-              bubbleSizeLevel={getBubbleLevel(timeMax, events.length)}
-              dividers={dividers}
-              bubbleLevelSizesInDivider={bubbleLevelSizesInDivider}
-              key={timeUnitKey + i}
-              onBubbleClick={() => {
-                updateLevel(level - 1, timeUnitKey)
-              }}
-              onSingleTimelineNodeSelect={() => {
-                setFocusUnitKey(timeUnitKey)
-              }}
-              isFocus={timeUnitKey === focusUnitKey}
-              headerHeight={headerHeight}
-              measure={measure}
+    measure !== 'event' ? (
+      <TimelineList
+        ref={timelineListRef}
+        timeUnitKeys={timeUnitKeysToRender}
+        timeUnitEvents={timeUnitEvents}
+        timeMax={timeMax}
+        divider={divider}
+        bubbleLevelSizesInDivider={bubbleLevelSizesInDivider}
+        onBubbleClick={(timeUnitKey) => {
+          updateLevel(level - 1, timeUnitKey)
+        }}
+        onSingleTimelineNodeSelect={(timeUnitKey) => {
+          setFocusUnitKey(timeUnitKey)
+        }}
+        focusUnitKey={focusUnitKey}
+        headerHeight={headerHeight}
+        measure={measure}
+        firstTimeUnitKey={firstTimeUnitKeyToRender}
+        lastTimeUnitKey={lastTimeUnitKeyToRender}
+        listDimension={listDimension}
+        listItemHeight={listItemHeight}
+        onTimelineListScroll={onTimelineListScroll}
+        isFixedMode={isFixedMode}
+      />
+    ) : (
+      timeUnitKeysToRender.map((timeUnitKey) => {
+        const event = timeUnitEvents[timeUnitKey]
+        return (
+          <EventWrapper key={timeUnitKey} id={`node-${timeUnitKey}`}>
+            <TimelineEvent
+              event={event}
+              fetchImageBaseUrl={fetchImageBaseUrl}
               timeUnitKey={timeUnitKey}
-              isTheFirstOrLastUnit={
-                timeUnitKey === firstTimeUnitKey ||
-                timeUnitKey === lastTimeeUnitKey
-              }
             />
-          )
-        })
-      : timeUnitKeysToRender.map((timeUnitKey) => {
-          const event = timeUnitEvents[timeUnitKey]
-          return (
-            <EventWrapper key={timeUnitKey} id={`node-${timeUnitKey}`}>
-              <TimelineEvent
-                event={event}
-                fetchImageBaseUrl={fetchImageBaseUrl}
-                timeUnitKey={timeUnitKey}
-              />
-            </EventWrapper>
-          )
-        })
+          </EventWrapper>
+        )
+      })
+    )
   return (
     <TagsContext.Provider value={{ tags, addTag, removeTag }}>
-      <Wrapper>
+      {isFixedMode && <BlockBodyScrollGlobalStyle />}
+      <Wrapper
+        ref={wrapperRef}
+        headerHeight={headerHeight}
+        isFixedMode={isFixedMode}
+      >
         <TimelineWrapper>
-          <div id="top" ref={topRef} />
-          <GlobalStyles />
           <TimelineNodesWrapper
+            id="containerRef"
             ref={containerRef}
             eventMode={measure === 'event'}
+            height={listDimension.height}
+            isFixedMode={isFixedMode}
           >
             {timelineNodesJsx}
           </TimelineNodesWrapper>
@@ -392,7 +464,6 @@ export default function Timeline({
             maxLevel={maxLevel}
             level={level}
             updateLevel={updateLevel}
-            stickyStrategy={stickyStrategy}
             selectedTags={tags}
             allTags={allTags}
             headerHeight={headerHeight}
@@ -401,7 +472,6 @@ export default function Timeline({
             <TimelineEventPanel
               event={focusEvent}
               fetchImageBaseUrl={fetchImageBaseUrl}
-              stickyStrategy={stickyStrategy}
               timeUnitKey={focusUnitKey}
               headerHeight={headerHeight}
               timeUnitKeys={timeUnitKeys}
@@ -413,7 +483,6 @@ export default function Timeline({
               }}
             />
           )}
-          <div id="bottom" ref={bottomRef} />
         </TimelineWrapper>
       </Wrapper>
     </TagsContext.Provider>